Added testing .competitions

+++ backend/src/api/internal/auth/domain/services.py
@@ -35,10 +35,10 @@ class Payload:
     EXPIRES_IN = "expires_in"
     PERMISSION = "permission"
 
-    def __init__(self, token_type: str, expires_in: int, user_id: int, permission: int):
+    def __init__(self, token_type: str, expires_in: float, user_id: int, permission: int):
         self._token_type = token_type
         self._user_id = user_id
-        self._expires_in = expires_in
+        self._expires_in = int(expires_in)
         self._permission = permission
 
     @property
